requirements.txt
pip install 

transformers
datasets
deepspeed
pyyaml
accelerate
